+++ lib/Poloniex/Client.rb
@@ -1,8 +1,0 @@
-# Poloniex/Client.rb
-# Poloniex::Client
-
-require_relative './V1/Client'
-
-module Poloniex
-  class Client < Poloniex::V1::Client; end
-end