+++ test/helper.rb
@@ -4,23 +4,22 @@ require 'minitest-spec-context'
 require 'vcr'
 require 'webmock'
 
-$LOAD_PATH.unshift(File.expand_path('../lib', __dir__))
-require 'poloniex'
+require_relative '../lib/poloniex'
 
 VCR.configure do |config|
   config.cassette_library_dir = 'test/fixtures/vcr_cassettes'
 
   config.hook_into :webmock
 
-  config.filter_sensitive_data('<API_KEY>') { 'test_api_key' }
-  config.filter_sensitive_data('<API_SECRET>') { 'test_api_secret' }
+  config.filter_sensitive_data('<API_KEY>'){ENV['POLONIEX_API_KEY']}
+  config.filter_sensitive_data('<API_SECRET>'){ENV['POLONIEX_API_SECRET'] }
 
   # Allow localhost connections for debugging
   config.ignore_localhost = true
 
-  config.default_cassette_options = { 
+  config.default_cassette_options = {
     record: :new_episodes,
-    match_requests_on: [:method, :uri]
+    match_requests_on: [:method, :path, :body]
   }
 end