+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.0] - 2025-07-10
+- + account endpoints
+- + margin endpoints
+- + trading endpoints
+- + wallet endpoints
+
 ## [0.0.0] - 2025-07-02
 - Initial release
 - Support for public REST API endpoints